bug fixed.

+++ cpymo/cpymo_audio.c
@@ -198,8 +198,8 @@ static void cpymo_audio_mix_samples(
 	len /= sizeof(TYPE); \
 	\
 	for (size_t i = 0; i < len; ++i) { \
-		float src_sample = (float)src[i] / (float)TYPE_MAX_VAL; \
-		float dst_sample = (float)dst[i] / (float)TYPE_MAX_VAL; \
+		double src_sample = (float)src[i] / (float)TYPE_MAX_VAL; \
+		double dst_sample = (float)dst[i] / (float)TYPE_MAX_VAL; \
 		src_sample *= volume * 1.0f; \
 		dst_sample += src_sample; \
 		\
